[SPARK-29412][SQL] refine the document of v2 session catalog config
Refine the document of v2 session catalog config, to clearly explain what it is, when it should be used and how to implement it.

Make this config more understandable

No

Pass the Jenkins with the newly updated test cases.

Closes #26071 from cloud-fan/config.

Authored-by: Wenchen Fan <wenchen@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>
(cherry picked from commit 9407fba0375675d6ee6461253f3b8230e8d67509)
Signed-off-by: colinma <colinma@tencent.com>